requirements.txt
pandas
numpy
matplotlib
seaborn
yfinance
python-pptx
